create generic paginated fetcher, moved objToQuery function into this fetcher
